requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
schedule>=1.2.0
questionary>=2.0.1
python-dotenv>=1.0.0
psutil>=5.9.8
feedparser>=6.0.11
cryptography>=42.0.0
keyring>=25.0.0
openai>=1.12.0
anthropic>=0.18.1
google-generativeai>=0.4.0